PENDING: arm: mach-k3: common: Fix warnings in lpm startup code
Fix the build time warnings generated from lpm code section.

Fixes warnings introduced in
commit 8a1f6eeb032 ("PENDING: arm: mach-k3: am62p5_init: Resume on LPM exit")
commit d87c6474214 ("PENDING: arm: mach-k3: am62a7_init: Resume on LPM exit")
commit 84e4c76124d ("PENDING: arm: mach-k3: common: Add lpm_resume_from_ddr")

Cc: Markus Schneider-Pargmann <msp@baylibre.com>
Tested-by: Kendall Willis <k-willis@ti.com>

+++ arch/arm/mach-k3/am62ax/am62a7_init.c
@@ -194,7 +194,7 @@ void board_init_f(ulong dummy)
 		panic("DRAM init failed: %d\n", ret);
 
 	if (wkup_ctrl_is_lpm_exit()) {
-		u64 meta_data_addr;
+		u32 meta_data_addr;
 
 		ret = wkup_r5f_am62_lpm_meta_data_addr(&meta_data_addr);
 		if (ret)
