+++ arch/arm/mach-k3/am62px/am62p5_init.c
@@ -198,7 +198,7 @@ void board_init_f(ulong dummy)
 	}
 
 	if (wkup_ctrl_is_lpm_exit()) {
-		u64 meta_data_addr;
+		u32 meta_data_addr;
 
 		ret = wkup_r5f_am62_lpm_meta_data_addr(&meta_data_addr);
 		if (ret)
